+++ Credits.lua
@@ -27,6 +27,7 @@ local donations = {
     {Name = "Bur", Realm = "Frostmane EU", Class = "DRUID"},
     {Name = "Ta", Realm = "Der Rat von Dalaran EU", Class = "SHAMAN"},
     {Name = "Estellar", Realm = "Bronze Dragonflight EU", Class = "PALADIN"},
+    {Name = "Seby", Realm = "TwistingNether EU", Class = "HUNTER"},
 }
 
 function credits.GetDonationsAsString()