+++ Localization/deDE.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-07-13 20-42-14 ]] --
+-- [[ Exported at 2022-07-14 09-04-58 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --